fix: fix webpack warns and prepare to remove karma
