+++ src/config/karma.conf.js
@@ -72,6 +72,10 @@ const karmaConfig = (config, argv) => {
       }
     ]),
 
+    proxies: {
+      '/test/fixtures/': '/base/test/fixtures/'
+    },
+
     preprocessors: files.reduce((acc, f) => {
       acc[f] = ['webpack', 'sourcemap']
       return acc
